+++ src/main/kotlin-go-support/runconfig/GocBuildOptions.kt
@@ -1,0 +1,10 @@
+package com.github.anniext.gocview.runconfig
+
+import com.intellij.execution.configurations.RunConfigurationOptions
+
+/**
+ * Goc 构建选项配置
+ *
+ * 用于存储 Goc Build 运行配置的选项
+ */
+class GocBuildOptions : RunConfigurationOptions()